refactor(textlint-formatter): modify test cases
