+++ packages/textlint-formatter/test/formatters/compact.js
@@ -4,13 +4,13 @@
  */
 
 "use strict";
+import formatter from "../../src/formatters/compact";
 
 //------------------------------------------------------------------------------
 // Requirements
 //------------------------------------------------------------------------------
 
-var assert = require("power-assert"),
-    formatter = require("../../lib/formatters/compact");
+var assert = require("power-assert");
 
 //------------------------------------------------------------------------------
 // Tests
