+++ packages/textlint-formatter/test/formatters/json.js
@@ -1,17 +1,17 @@
 /**
-  * @fileoverview Tests for JSON reporter.
+ * @fileoverview Tests for JSON reporter.
  * @author Burak Yigit Kaya aka BYK
  * @copyright 2015 Burak Yigit Kaya. All rights reserved.
  */
 
 "use strict";
+import formatter from "../../src/formatters/json";
 
 //------------------------------------------------------------------------------
 // Requirements
 //------------------------------------------------------------------------------
 
-var assert = require("power-assert"),
-    formatter = require("../../lib/formatters/json");
+var assert = require("power-assert");
 
 //------------------------------------------------------------------------------
 // Tests
