+++ packages/textlint-formatter/test/formatters/pretty-error-test.js
@@ -1,8 +1,9 @@
 // LICENSE : MIT
 "use strict";
+import prettyError from "../../src/formatters/pretty-error";
+
 var assert = require("power-assert");
 var path = require("path");
-var prettyError = require("../../lib/formatters/pretty-error");
 var stripAnsi = require("strip-ansi");
 describe("pretty-error", function() {
     context("when first line", function() {
