+++ packages/textlint-formatter/test/tsconfig.json
@@ -1,0 +1,17 @@
+{
+  "extends": "../tsconfig.json",
+  "compilerOptions": {
+    "outDir": "../out",
+    "allowJs": true,
+    "declaration": false,
+    "sourceMap": true
+  },
+  "include": [
+    "../src/**/*",
+    "./**/*"
+  ],
+  "exclude": [
+    ".git",
+    "node_modules"
+  ]
+}
